tests/wpt/mozilla/tests/mozilla/mozbrowser/mozbrowsersecuritychange_event.html

<head>
  <title>mozbrowsersecuritychange event is dispatched when content security state changes</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>

    async_test(function(t) {

      const HTTP_URL = "http://web-platform.test:8000";
      const HTTPS_URL = "https://web-platform.test:8443";

      var urls = [ HTTP_URL, HTTPS_URL, HTTP_URL ];
      var url_index = 0;

      var expectedEvents = [
        "insecure",
        "secure",
        "insecure",
      ];

      var receivedEvents = [];

      var iframe = document.createElement("iframe");
      iframe.mozbrowser = "true";
      iframe.src = urls[url_index++];

      iframe.addEventListener("mozbrowsersecuritychange", t.step_func(e => {

        receivedEvents.push(e.detail.state);

        if (receivedEvents.length == expectedEvents.length) {
          assert_array_equals(receivedEvents, expectedEvents);
          t.done();
        } else {
          iframe.src = urls[url_index++];
        }
      }));


      document.body.appendChild(iframe);
    });

  </script>
</body>
